Add useInsertPreset helper.
